Implemented pipline, module classes
