+++ main/modules/module.py
@@ -1,0 +1,6 @@
+import queue
+
+class Module:
+    def __init__(self):
+        self.input = queue.Queue()
+        self.output = queue.Queue()
